issue#01: New order. Creation of test

+++ src/main/java/es/upm/miw/apaw/ecp2/api/entities/Professional.java
@@ -24,7 +24,7 @@ public class Professional {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(long phone) {
         this.phone = phone;
     }
